add SingleColorChannelWithAlphaMapper to handle mapping (transforming) color sources with masks; update expected ArgbRendererTest result images to accommodate different jpg compression of ByteProcessor sources now that everything is not converted to a FloatProcessor first; ShortRendererTest is still broken and will be fixed shortly :)

+++ render-app/src/main/java/org/janelia/alignment/mapper/SingleChannelWithAlphaMapper.java
@@ -58,8 +58,8 @@ public class SingleChannelWithAlphaMapper
 
     public void setBlendedIntensity(final int targetX,
                                     final int targetY,
-                                    final double sourceIntensity,
-                                    final double sourceMaskIntensity) {
+                                    final int sourceIntensity,
+                                    final int sourceMaskIntensity) {
 
         final double sourceAlpha = sourceMaskIntensity / sourceMaxMaskIntensity;
         final double targetIntensity = target.ip.getf(targetX, targetY);
